pyproject.toml
[tool.poetry]
name = "US-funds-project"
version = "0.1.0"
description = ""
authors = ["Carla Teixeira Coutinho"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "2.0.3"
dbt-duckdb = "^1.7.1"
dbt-core = "^1.7.8"
apache-superset = "^3.1.0"
sqlfluff = "^2.3.5"
streamlit = "^1.31.1"
matplotlib = "^3.8.3"
plotly = "^5.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.sqlfluff.core]
dialect = "duckdb"
templater = "jinja"
sql_file_exts = ".sql,.sql.j2,.dml,.ddl"

[tool.sqlfluff.indentation]
tab_space_size = 4
